# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Druglane Cloud API",
        "description": "Cloud-based management system for offline desktop inventory management applications. This API handles license management, backup storage, and provides online access to inventory data.",
        "contact": {
            "email": "admin@druglane.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/company-data/products": {
            "get": {
                "tags": [
                    "Company Data"
                ],
                "summary": "Get company products",
                "description": "Retrieve paginated list of products from the company's SQLite database. Company users only.",
                "operationId": "3863df3057232880fff005a14a31a684",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page offset (not page number)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 0
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "example": 100
                        }
                    },
                    {
                        "name": "param",
                        "in": "query",
                        "description": "Search query for product name, barcode, or category",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Paracetamol"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Paracetamol 500mg"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "Analgesics"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 2.5
                                                    },
                                                    "cost_price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1.5
                                                    },
                                                    "current_stock": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 150
                                                    },
                                                    "min_stock": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 20
                                                    },
                                                    "expiry": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    },
                                                    "barcode": {
                                                        "type": "string",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 250
                                        },
                                        "displayColumns": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "columnLabels": {
                                            "type": "object"
                                        },
                                        "columnFilters": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No company associated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No company associated with this account"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Failed to retrieve products"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/company-data/sales": {
            "get": {
                "tags": [
                    "Company Data"
                ],
                "summary": "Get company sales",
                "description": "Retrieve paginated list of sales transactions from the company's SQLite database. Company users only.",
                "operationId": "3ab76eba394f7d6fc4c83a0273db3590",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page offset",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 0
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "example": 100
                        }
                    },
                    {
                        "name": "param",
                        "in": "query",
                        "description": "Search query for sale code, customer, or creditor name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "SALE-001"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Filter sales from this date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Filter sales until this date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-31"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "SALE-001"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2025-12-09"
                                                    },
                                                    "customer": {
                                                        "type": "string",
                                                        "example": "John Doe",
                                                        "nullable": true
                                                    },
                                                    "total_amount": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 125.5
                                                    },
                                                    "amount_paid": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 125.5
                                                    },
                                                    "payment_method": {
                                                        "type": "string",
                                                        "example": "Cash"
                                                    },
                                                    "credit_paid": {
                                                        "type": "integer",
                                                        "example": 0
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "displayColumns": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "columnLabels": {
                                            "type": "object"
                                        },
                                        "columnFilters": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "No company associated"
                    },
                    "500": {
                        "description": "Failed to retrieve sales"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}